requirements.txt
apache_sedona==1.8.1
geopandas==1.1.3
matplotlib==3.10.8
pandas==3.0.2
pyspark.egg==info
pyspark_client.egg==info
seaborn==0.13.2
 